Create addruleSG.py
This script was written to ingest a csv file with two columns — the first column contains the IP and the second contains a description of the rule. If your csv only has a column of IPs, you can easily remove the “description” key-value pair from IpRanges and remove or comment out the declaration of the the description variable in the for loop.
I wrote this to bulk add rules for the same port range — if you are adding rules for a variety of ports/port ranges, add two columns to your csv, one for the port range start value and one for the port range end value and populate the csv accordingly. Move the declaration of port_range_start and port_range_end to the for loop and add the appropriate row index.
Example:
for row in csv_f:
    cidr = row[0] + "/32"
    description = row [1]   
    port_range_start = row[2]
    port_range_end = row[3]